refactor(cli): extract costs command family into cli/commands/costs.ts
